+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/composable/AnimatedFlipperPalletProvider.kt
@@ -59,5 +59,6 @@ internal fun FlipperPallet.toAnimatePallet() = FlipperPallet(
     flipperDisableButton = animateColor(flipperDisableButton),
     borderViewReportBug = animateColor(borderViewReportBug),
 
-    screenStreamingBorderColor = animateColor(screenStreamingBorderColor)
+    screenStreamingBorderColor = animateColor(screenStreamingBorderColor),
+    screenStreamingNotConnectedColor = animateColor(screenStreamingNotConnectedColor)
 )