+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/dialogs/ComposableDialogPending.kt
@@ -3,6 +3,8 @@ package com.flipperdevices.updater.card.composable.dialogs
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.buildAnnotatedString
 import com.flipperdevices.core.ui.dialog.composable.FlipperDialog
 import com.flipperdevices.core.ui.dialog.composable.multichoice.FlipperMultiChoiceDialog
 import com.flipperdevices.core.ui.dialog.composable.multichoice.FlipperMultiChoiceDialogModel
@@ -29,15 +31,18 @@ fun FlipperDialogReadyUpdate(
         R.string.update_card_confirm_update
     }
 
-    val text = buildAnnotatedStringWithVersionColor(
-        version,
-        postfixId = R.string.update_card_confirm_desc
-    )
+    val description = buildAnnotatedString {
+        append(stringResource(R.string.update_card_confirm_desc_pre))
+        append(' ')
+        append(annotatedStringWithVersion(version))
+        append(' ')
+        append(stringResource(R.string.update_card_confirm_desc_post))
+    }
 
     val dialogModel = remember(onDismiss, onRun) {
         FlipperMultiChoiceDialogModel.Builder()
             .setTitle(titleId)
-            .setDescription(text)
+            .setDescription(description)
             .setOnDismissRequest(onDismiss)
             .addButton(buttonId, onRun, isActive = true)
             .addButton(R.string.update_card_confirm_stop, onDismiss)
@@ -47,14 +52,36 @@ fun FlipperDialogReadyUpdate(
 }
 
 @Composable
-fun FlipperDialogSynchronization(onCancel: () -> Unit, onPauseAndUpdate: () -> Unit) {
+fun FlipperDialogSynchronization(
+    isInstallUpdate: Boolean,
+    version: FirmwareVersion?,
+    onCancel: () -> Unit,
+    onPauseAndUpdate: () -> Unit
+) {
+    val buttonText = if (isInstallUpdate) {
+        R.string.update_card_dialog_pause_sync_install
+    } else {
+        R.string.update_card_dialog_pause_sync_update
+    }
+    val description = buildAnnotatedString {
+        if (isInstallUpdate) {
+            append(stringResource(R.string.update_card_dialog_pause_sync_desc_pre_install))
+        } else {
+            append(stringResource(R.string.update_card_dialog_pause_sync_desc_pre_update))
+        }
+        append(' ')
+        append(annotatedStringWithVersion(version))
+        append(' ')
+        append(stringResource(R.string.update_card_dialog_pause_sync_desc_post))
+    }
+
     val dialogModel = remember(onPauseAndUpdate, onCancel) {
         FlipperMultiChoiceDialogModel.Builder()
             .setTitle(R.string.update_card_dialog_pause_sync_title)
-            .setDescription(R.string.update_card_dialog_pause_sync_desc)
+            .setDescription(description)
             .setOnDismissRequest(onCancel)
             .addButton(
-                R.string.update_card_dialog_pause_sync_update,
+                buttonText,
                 onPauseAndUpdate,
                 isActive = true
             )